chore: adding hooks for use forums
