+++ modules/forums/index.tsx
@@ -1,11 +1,9 @@
-import axios from 'axios'
+import Title from './components/Title'
 import Layout from '@/components/layouts/main'
+import useForums from '@/hooks/useForums'
 import Discussion from './components/Discussion'
 import InputMessage from './components/InputMessage'
-import React, { useState, useEffect } from 'react'
-import { toast } from 'react-toastify'
 import { useSession } from 'next-auth/react'
-import Title from './components/Title'
 
 interface User {
   name: string
@@ -21,40 +19,11 @@ export interface Post {
 
 const Forums = () => {
   const { status } = useSession()
-  const [content, setContent] = useState<string>('')
-  const [posts, setPosts] = useState<Post[]>([])
-
-  const fetchPosts = async () => {
-    try {
-      const response = await axios.get<Post[]>('/api/forums')
-      setPosts(response.data)
-    } catch (err) {
-      console.error('Error fetching posts:', err)
-    }
-  }
-
-  useEffect(() => {
-    fetchPosts()
-  }, [])
+  const { posts, content, setContent, createPost } = useForums()
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault()
-
-    try {
-      await axios.post<Post>('/api/forums', {
-        content
-      })
-      toast.success('Message created successfully!', {
-        autoClose: 1000
-      })
-      setContent('')
-      fetchPosts()
-    } catch (err: any) {
-      console.error('Error:', err)
-      toast.error(err.response?.data?.message || 'Something went wrong', {
-        autoClose: 1000
-      })
-    }
+    await createPost()
   }
   return (
     <>
